Refactored VariableInfo into VariableAttributes. This change affected almost every class. Other major changes include new dialog for IrtItemCalibration and other IRT related changes.

+++ src/main/java/com/itemanalysis/jmetrik/commandbuilder/MegaOption.java
@@ -416,8 +416,8 @@ public class MegaOption{
         return Double.parseDouble(value);
     }
 
-    public int getValueAtAsInteger(String argName, double defaultValue)throws IllegalArgumentException{
-        String value = getValueAt(argName, Double.valueOf(defaultValue).toString());
+    public int getValueAtAsInteger(String argName, int defaultValue)throws IllegalArgumentException{
+        String value = getValueAt(argName, Integer.valueOf(defaultValue).toString());
         return Integer.parseInt(value);
     }
